tests/wpt/web-platform-tests/css/css-will-change/parsing/will-change-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Will Change Test: parsing will-change with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-will-change/#propdef-will-change">
<meta name="assert" content="will-change only supports the grammar 'auto | <animateable-feature>#'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("will-change", "auto transform");
test_invalid_value("will-change", "auto, transform");
test_invalid_value("will-change", "contents auto");
test_invalid_value("will-change", "contents, auto");

// CSS-wide keywords are excluded from <custom-ident>
// https://drafts.csswg.org/css-values-4/#identifier-value
test_invalid_value("will-change", "transform, initial");
test_invalid_value("will-change", "transform, inherit");
test_invalid_value("will-change", "transform, unset");
test_invalid_value("will-change", "transform, default");

// will-change additionally excludes the following from <custom-ident>
test_invalid_value("will-change", "will-change");
test_invalid_value("will-change", "none");
test_invalid_value("will-change", "all");
</script>
</body>
</html>
